requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
coverage==7.6.1
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==65.0.2
django-rest-knox==5.0.2
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
factory_boy==3.3.1
Faker==30.1.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
